Split dashboard workspaces into focused modules
