+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/token-action-card.tsx
@@ -1,0 +1,22 @@
+"use client";
+
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
+import type { ReactNode } from "react";
+
+interface TokenActionCardProps {
+  title: string;
+  description?: string;
+  children: ReactNode;
+}
+
+export function TokenActionCard({ title, description, children }: TokenActionCardProps) {
+  return (
+    <Card className="gap-4">
+      <CardHeader>
+        <CardTitle>{title}</CardTitle>
+        {description ? <CardDescription>{description}</CardDescription> : null}
+      </CardHeader>
+      <CardContent className="space-y-4">{children}</CardContent>
+    </Card>
+  );
+}
