+++ apps/sdp-web/src/app/dashboard/issuance/create-token-modal.types.ts
@@ -1,0 +1,39 @@
+export type TemplateSelection = "stablecoin" | "custom" | "tokenized-security" | "arcade";
+export type CreationStep = "identity" | "features";
+export type AccessControlMode = "allowlist" | "blocklist";
+
+export type FlowState =
+  | {
+      kind: "templateSelection";
+    }
+  | {
+      kind: "creation";
+      template: TemplateSelection;
+      step: CreationStep;
+    };
+
+export interface TokenDraft {
+  template: TemplateSelection | null;
+  uri: string;
+  name: string;
+  symbol: string;
+  decimals: "" | "0" | "6" | "8" | "9";
+  accessControlMode: AccessControlMode;
+}
+
+export interface TemplateCardDescriptor {
+  id: string;
+  name: string;
+  description: string;
+  iconClassName: string;
+  enabled: boolean;
+  template?: TemplateSelection;
+}
+
+export interface IdentityValidation {
+  uriValid: boolean;
+  nameValid: boolean;
+  symbolValid: boolean;
+  decimalsValid: boolean;
+  isValid: boolean;
+}
