+++ apps/sdp-web/src/app/dashboard/payments/payments-workspace.types.ts
@@ -1,0 +1,7 @@
+import type { ComplianceProviderResult } from "@/lib/compliance";
+
+export interface ComplianceSnapshot {
+  address: string;
+  checkedAt: string;
+  providers: ComplianceProviderResult[];
+}
